requirements.txt
pillow==11.0.0
pycups==2.0.4
